% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chopIntervals.R
\name{chopIntervals}
\alias{chopIntervals}
\title{"Chopping" overlapping intervals in dataframe}
\usage{
chopIntervals(data, id, start.col, end.col)
}
\arguments{
\item{data}{R data.frame or tibble}

\item{id}{Id variable (function assumes grouped values)}

\item{start}{col Variable defining end of interval}
}
\value{
Dataframe containing series
}
\description{
Function to chop overlapping ranges (date or integer) into non-overlapping ones
}
