% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGroupedNAs.R
\name{plotGroupedNAs}
\alias{plotGroupedNAs}
\title{Plot missing values in dataframe by grouping variable}
\usage{
plotGroupedNAs(data, groupvar)
}
\arguments{
\item{data}{R data.frame or tibble}

\item{groupvar}{Variable to group missing observations by}
}
\value{
\code{ggplot2} plot
}
\description{
Returns a tile plot with percentages of missing observations for each grouped variable
}
