% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotNAs.R
\name{plotNAs}
\alias{plotNAs}
\title{Plot missing values in dataframe}
\usage{
plotNAs(data)
}
\arguments{
\item{data}{dataframe}
}
\value{
\code{ggplot2} plot
}
\description{
Returns a bar plot with missing values for each variable
}
