% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unionDFs.R
\name{unionDFs}
\alias{unionDFs}
\title{Union dataframes with non-matching column-names / numbers}
\usage{
unionDFs(df1, df2, cols1, cols2)
}
\arguments{
\item{cols1}{Columns of data.frame 1 to be included}

\item{cols2}{Columns of data.frame 2 to be included}

\item{data1}{R data.frame or tibble 1}

\item{data2}{R data.frame or tibble 2}
}
\value{
Union of dataframe 1 and 2 for specified columns
}
\description{
Union dataframes with non-matching column-names / numbers
}
